Move corners function used in diagnostics to module
