+++ diagnostics/physics/ssh_eval.py
@@ -17,7 +17,7 @@ import logging
 
 from cefi_kit.io import load_config, open_var
 from cefi_kit.plots import add_ticks, annotate_skill, autoextend_colorbar, get_map_norm, save_figure
-from plot_common import corners
+from cefi_kit.grids import corners
 
 # Configure logging for ssh_eval
 logger = logging.getLogger(__name__)
